nginx: and nginx-util: remove nossl variant
For nginx leave only the ssl and the all-module variant.
For nginx-util leave the ssl and the ssl-nopcre versions only.
Replace the removed packages by dummies for a smooth upgrade.
The dummy packages will be removed after a transition period.

Additionally, enable the http2 module for nginx-ssl by default
and cleanup the nginx.init file.

Signed-off-by: Peter Stadler <peter.stadler@student.uibk.ac.at>